Small stability improvements
1. Add timeout to javacOptions request, as SBT hangs in case it's called

2. addTargetToSetIfNoneOfItsNeighborsIsAddedAndDoTheSameForDependencies method
is recursive, but it was terminated only when target.dependencies was an empty list,
as this is only case when `fold`'s inner method (the `addTarget...Dependencies` method itself).
In case when dependency graph is not acyclic, this will be called infinitely, that's why we need
to care we don't process graph's nodes twice. That's why we skip the ones that has already been
added to `nonOverlappingTargetsGraph`.

3. inherit stderr from bsp process, just to make sure it does not hang if buffer is full