[fix] avoid misinterpretation of dots in target names | #BAZEL-391 Done
Previously, targets //A/B:C, //A:B.C and //A.B:C would obtain similar module name (A.B.C), thus creating a clash. After this change it would be A.B.C, A.B:C and A:B.C respectively. The : character has been chosen for dots' replacement, because it's not allowed in Bazel labels (https://bazel.build/concepts/labels#target-names), so it will not cause further clashes.

Merge-request: BAZEL-MR-517
Merged-by: Marcin Kocot <Marcin.Kocot@jetbrains.com>